Put the different model in the dtd. Permit the xml validation before parsing.

+++ src/surf/surfxml_parse.c
@@ -556,7 +556,7 @@ void STag_surfxml_trace_connect(void){
   surfxml_call_cb_functions(STag_surfxml_trace_connect_cb_list);
 }
 void STag_surfxml_AS(void){
-  sg_platf_new_AS_begin(A_surfxml_AS_id,A_surfxml_AS_routing);
+  sg_platf_new_AS_begin(A_surfxml_AS_id, (int)A_surfxml_AS_routing);
 }
 void ETag_surfxml_AS(void){
   sg_platf_new_AS_end();
